+++ src/monitoring/monitoring.service.ts
@@ -5,34 +5,8 @@
  */
 
 import { Injectable } from '@nestjs/common';
-import { promises as fs } from 'fs';
-import { join as joinPath } from 'path';
-import { ServerStatusDto, ServerVersion } from './server-status.dto';
-
-let versionCache: ServerVersion;
-
-async function getServerVersionFromPackageJson(): Promise<ServerVersion> {
-  if (versionCache === null) {
-    const rawFileContent: string = await fs.readFile(
-      joinPath(__dirname, '../../package.json'),
-      { encoding: 'utf8' },
-    );
-    // TODO: Should this be validated in more detail?
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-    const packageInfo: { version: string } = JSON.parse(rawFileContent);
-    const versionParts: number[] = packageInfo.version
-      .split('.')
-      .map((x) => parseInt(x, 10));
-    versionCache = {
-      major: versionParts[0],
-      minor: versionParts[1],
-      patch: versionParts[2],
-      preRelease: 'dev', // TODO: Replace this?
-    };
-  }
-
-  return versionCache;
-}
+import { ServerStatusDto } from './server-status.dto';
+import { getServerVersionFromPackageJson } from '../utils/serverVersion';
 
 @Injectable()
 export class MonitoringService {
